pyproject.toml
[tool.poetry]
name = "similarity-service"
version = "0.1.0"
description = ""
authors = ["akroosh <antonkroosh@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
grpcio-tools = "^1.56.0"
SQLAlchemy = "^2.0.18"
alembic = "^1.11.1"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
nltk = "^3.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
